+++ src/context/datasetContext/index.test.tsx
@@ -155,9 +155,59 @@ describe("Testing DatasetContext", () => {
       error: undefined,
     });
     documentBody = render(
+      <DatasetProvider datasetUrl="https://some-broken-resource.com">
+        <ExampleComponentWithDataset />
+      </DatasetProvider>
+    );
+
+    const { baseElement } = documentBody;
+    expect(baseElement).toMatchSnapshot();
+  });
+
+  it("matches snapshot when fetching and loadingComponent is passed", async () => {
+    (useDataset as jest.Mock).mockReturnValue({
+      solidDataset: undefined,
+      error: undefined,
+    });
+    documentBody = render(
       <DatasetProvider
         datasetUrl="https://some-broken-resource.com"
-        loading={() => "loading"}
+        loadingComponent={() => <span>loading component</span>}
+      >
+        <ExampleComponentWithDataset />
+      </DatasetProvider>
+    );
+
+    const { baseElement } = documentBody;
+    expect(baseElement).toMatchSnapshot();
+  });
+
+  it("matches snapshot when fetching and loading is passed", async () => {
+    (useDataset as jest.Mock).mockReturnValue({
+      solidDataset: undefined,
+      error: undefined,
+    });
+    documentBody = render(
+      <DatasetProvider
+        datasetUrl="https://some-broken-resource.com"
+        loading={() => <span>loading</span>}
+      >
+        <ExampleComponentWithDataset />
+      </DatasetProvider>
+    );
+
+    const { baseElement } = documentBody;
+    expect(baseElement).toMatchSnapshot();
+  });
+  it("matches snapshot when loadingComponent is null", async () => {
+    (useDataset as jest.Mock).mockReturnValue({
+      solidDataset: undefined,
+      error: undefined,
+    });
+    documentBody = render(
+      <DatasetProvider
+        datasetUrl="https://some-broken-resource.com"
+        loadingComponent={null}
       >
         <ExampleComponentWithDataset />
       </DatasetProvider>